feat(S11P21D109-284): 거래내역 구현

See merge request s11-blockchain-nft-sub1/S11P21D109!135

+++ backend/spring-server/src/main/java/com/semonemo/spring_server/domain/coin/dto/request/CoinRequestDto.java
@@ -13,5 +13,5 @@ import java.util.List;
 @NoArgsConstructor
 @ToString
 public class CoinRequestDto {
-    private BigInteger amount;
+    private Long amount;
 }
